Continuation of Brian's work bpkroth/more-cpp-compiler-warning-fixups.
Enable clang compilation flags: -Wall -Wextra -Wpedantic -Werror.
Remove macros from Mlos.Core.
Rename variable name from elementLength to elementDataSize.

+++ source/Examples/SmartCache/SmartCacheImpl.h
@@ -19,7 +19,7 @@ template<typename TKey, typename TValue>
 class SmartCacheImpl
 {
 private:
-    int m_cacheSize;
+    size_t m_cacheSize;
 
     std::list<TValue> m_elementSequence;
